project_3_(chat_with_pdf)/requirements.txt
gradio 
openai 
PyPDF2 
python-dotenv